django/requirements.txt
# ── Web framework ──────────────────────────────
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2          # OpenAPI schema

# ── Database ───────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pgvector==0.2.5
django-pgvector==0.3.0

# ── Async / Tasks ──────────────────────────────
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
kombu==5.3.4

# ── Search ─────────────────────────────────────
elasticsearch==8.13.0
elasticsearch-dsl==8.9.0

# ── AI / ML ────────────────────────────────────
openai==1.25.0
anthropic==0.25.3
sentence-transformers==2.7.0
spacy==3.7.4
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# ── HTTP / Scraping ────────────────────────────
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

# ── Auth / Security ────────────────────────────
PyJWT==2.8.0
cryptography==42.0.5
django-allauth==0.61.1

# ── Utilities ──────────────────────────────────
python-dotenv==1.0.1
Pillow==10.3.0
gunicorn==22.0.0
whitenoise==6.6.0
django-storages==1.14.3

# ── Email ──────────────────────────────────────
sendgrid==6.11.0

# ── Dev tools ──────────────────────────────────
ipython==8.24.0
django-extensions==3.2.3
